+++ packages/flutter_chat_view/lib/src/screens/chat_screen.dart
@@ -20,7 +20,7 @@ class ChatScreen extends StatefulWidget {
     this.unreadMessageTextStyle,
     this.onNoChats,
     this.deleteChatDialog,
-    this.translations = const ChatTranslations(),
+    this.translations = const ChatTranslations.empty(),
     this.disableDismissForPermanentChats = false,
     super.key,
   });
@@ -119,7 +119,8 @@ class _ChatScreenState extends State<ChatScreen> {
             child: ListView(
               controller: controller,
               physics: const AlwaysScrollableScrollPhysics(),
-              padding: const EdgeInsets.fromLTRB(28, 16, 28, 0),
+              padding: widget.options.paddingAroundChatList ??
+                  const EdgeInsets.fromLTRB(28, 16, 28, 0),
               children: [
                 StreamBuilder<List<ChatModel>>(
                   stream: widget.service.chatOverviewService.getChatsStream(),
@@ -384,7 +385,6 @@ class ChatListItem extends StatelessWidget {
               ),
             ),
           ),
-          const Divider(),
         ],
       );
 }